requirements.txt
python-vlc
async-property
python-telegram-bot>=20.0a0
aiosqlite
